Lint, rename some things to use better terminology for the new graphql endpoint
